Restructured project to clean up document tree
